service/add/validation á eftir að setja i foll

+++ genius.cpp
@@ -5,7 +5,7 @@ Genius::Genius()
 
 }
 
-Genius::Genius(string name, string gender, size_t year_of_birth, size_t year_of_death)
+Genius::Genius(string name, string gender, unsigned int year_of_birth, unsigned int year_of_death)
 {
     _name = name;
     _gender = gender;
