+++ servicelayer.h
@@ -15,7 +15,7 @@ public:
     vector<Genius> sortVector();
     Genius find(string name) const;
     vector<Genius> getGenius();
-    bool addEntry(string name, char gender, size_t date_of_birth, size_t date_of_death);
+    bool addEntry(string name, char gender, unsigned int date_of_birth, unsigned int date_of_death);
 };
 
 #endif SERVICELAYER_H
